+++ src/main/java/roomescape/repository/JdbcReservationTimeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.infrastructure;
+package roomescape.repository;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
